requirements.txt
# Deep Learning & Computer Vision
# Requires Python 3.9+
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
mediapipe>=0.8.9

# GUI
customtkinter>=5.2.0
Pillow>=9.0.0

# OCR
pytesseract>=0.3.10

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# YOLOv5 Dependencies
matplotlib>=3.2.2
scipy>=1.4.1
tqdm>=4.64.0
seaborn>=0.11.0
PyYAML>=5.3.1
requests>=2.23.0
tensorboard>=2.4.1
thop
